+++ integration/test_qsys_jobs.py
@@ -4,14 +4,14 @@ import os
 from datetime import datetime
 from typing import Any, cast
 
-import pytest
 from guppylang import guppy  # type: ignore
+from guppylang.qsys_result import QsysResult
 from guppylang.std.builtins import result
 from guppylang.std.quantum import cx, h, measure, qubit, x, z
 from pytket.backends.backendinfo import BackendInfo
-from quantinuum_schemas.models.result import QSysResult
 
 import qnexus as qnx
+from qnexus.models.references import HUGRRef
 
 
 def prepare_teleportation() -> Any:
@@ -43,9 +43,6 @@ def prepare_teleportation() -> Any:
     return main.compile()
 
 
-@pytest.mark.skip(
-    "Skipping QSys tests until we have a consistent test device to target"
-)
 def test_guppy_execution(
     _authenticated_nexus: None,
     qa_project_name: str,
@@ -76,7 +73,6 @@ def test_guppy_execution(
         name=f"QA Test QSys job from {datetime.now()}",
     )
 
-    # QSYS QA device might not always be online, so we might expect failures for now
     qnx.jobs.wait_for(job_ref)
 
     results = qnx.jobs.results(job_ref)
@@ -85,11 +81,14 @@ def test_guppy_execution(
     result_ref = results[0]
 
     assert isinstance(result_ref.download_backend_info(), BackendInfo)
-    assert isinstance(result_ref.get_input(), hugr_package)
+    assert isinstance(result_ref.get_input(), HUGRRef)
 
     assert result_ref.get_input().id == hugr_ref.id
 
-    qsys_result = cast(QSysResult, result_ref.download_result())
-    assert len(qsys_result) == n_shots
-    assert qsys_result[0][0][0] == "teleported"
-    assert qsys_result[0][0][1] == 0
+    qsys_result = cast(QsysResult, result_ref.download_result())
+    assert len(qsys_result.results) == n_shots
+    assert qsys_result.results[0].entries[0][0] == "teleported"
+    assert qsys_result.results[0].entries[0][1] == 1
+
+    # check some QsysResults functionality
+    assert len(qsys_result.collated_counts().items()) > 0